+++ src/SelectPipe.ts
@@ -1,0 +1,8 @@
+import { Pipe } from '@angular/core';
+
+import { ApolloQueryPipe } from './ApolloQueryPipe';
+
+@Pipe({
+  name: 'select',
+})
+export class SelectPipe extends ApolloQueryPipe {}