+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Made queries reusable (use refetch on new variables) ([PR #74](https://github.com/apollostack/angular2-apollo/pull/74))
 - Used [`apollo-client-rxjs`](https://github.com/kamilkisiela/apollo-client-rxjs) ([PR #72](https://github.com/apollostack/angular2-apollo/pull/72))
 - Fixed an issue that prevents from subscribing to `ApolloQueryObservable` ([PR #71](https://github.com/apollostack/angular2-apollo/pull/71))
+- Added `SelectPipe` and deprecated `ApolloQueryPipe` ([PR #78](https://github.com/apollostack/angular2-apollo/pull/78))
 
 ### v0.4.3